Make a more specific exception
I always wrap my RNCryptor code in a try, except clause, because bad data is somewhat commonplace in my program. It would be nice to be 100% certain that the exception is caused by bad data and handle accordingly, rather than the exception being handled as if it were bad data when it could be a variety of other problems. With this, I can catch the DecryptionError instead of any generic Exception.